+++ Vehicles/Wheel.hpp
@@ -7,10 +7,10 @@ namespace Vehicles
 	{
 		
 	public:
-		Wheel(double diameter);		// declare constructor (you must provide a diameter to create a wheel)
+		Wheel(int diameter);		// declare constructor (you must provide a diameter to create a wheel)
 		std::string WheelReport();  // declare a function that will tell us about the wheel
 	private:
-		double m_diameter;			// declare a private variable to contain the diameter
+		int m_diameter;			// declare a private variable to contain the diameter
 	};
 
 }
